Change for a64 build

+++ build.sh
@@ -13,15 +13,48 @@ set -e
 BL=$PWD/treble_build_pe
 BD=$HOME/builds
 BRANCH=$1
+export WITH_SU=false
+export USE_CCACHE=1
+export CCACHE_COMPRESS=1
+export CCACHE_MAXSIZE=50G
+export BUILD_USER=crazyads69
+export BUILD_HOST=crazyads69
+export BUILD_USERNAME=crazyads69
+export BUILD_HOSTNAME=crazyads69
+export KBUILD_BUILD_USER=crazyads69
+export KBUILD_BUILD_HOST=crazyads69
 
 [ "$BRANCH" == "" ] && BRANCH="twelve"
 [ "$BRANCH" == "twelve" ] && BUILD="PixelExperience" || BUILD="PixelExperience_Plus"
 [ "$BRANCH" == "twelve" ] && PEMK="$BL/pe.mk" || PEMK="$BL/peplus.mk"
 
+installRequiredDependency() {
+    echo "---> Install Required Dependency"
+    apt install -y bc bison build-essential curl flex g++-multilib gcc-multilib git gnupg gperf libxml2 \
+                lib32z1-dev liblz4-tool libncurses5-dev libsdl1.2-dev libwxgtk3.0-gtk3-dev imagemagick git \
+                lunzip lzop schedtool squashfs-tools xsltproc zip zlib1g-dev openjdk-8-jdk python perl  \
+                xmlstarlet virtualenv xz-utils rr jq libncurses5 pngcrush lib32ncurses5-dev git-lfs libxml2 \
+                openjdk-11-jdk-headless repo
+
+    echo
+
+    apt install -y openjdk-8-jdk apache2 bc bison build-essential ccache curl \
+                flex g++-multilib gcc-multilib git gnupg gperf imagemagick lib32ncurses5-dev \
+                lib32readline-dev lib32z1-dev liblz4-tool libncurses5-dev libsdl1.2-dev \
+                libssl-dev libwxgtk3.0-dev libxml2 libxml2-utils lzop pngcrush rsync \
+                schedtool squashfs-tools xsltproc zip zlib1g-dev git-core gnupg flex \
+                bison gperf build-essential zip curl zlib1g-dev gcc-multilib g++-multilib \
+                libc6-dev-i386 lib32ncurses5-dev x11proto-core-dev libx11-dev lib32z-dev ccache \
+                libgl1-mesa-dev libxml2-utils xsltproc unzip python python3 libncurses5 repo
+
+    echo
+
+}
+
 initRepos() {
     if [ ! -d .repo ]; then
         echo "--> Initializing PE workspace"
-        repo init -u https://github.com/PixelExperience/manifest -b $BRANCH
+        repo init -u https://github.com/PixelExperience/manifest -b $BRANCH --groups=all,notdefault,default,-device,-darwin,-x86,-mips,-exynos5,-mako,-lge,-coral,-goldfish --depth=1
         echo
 
         echo "--> Preparing local manifest"
@@ -33,7 +66,7 @@ initRepos() {
 
 syncRepos() {
     echo "--> Syncing repos"
-    repo sync -c --force-sync --no-clone-bundle --no-tags -j$(nproc --all)
+    repo sync -c --force-sync --no-clone-bundle --no-tags -j$(nproc --all) --optimized-fetch --prune
     echo
 }
 
@@ -66,45 +99,40 @@ buildTrebleApp() {
     echo "--> Building treble_app"
     cd treble_app
     bash build.sh release
-    cp TrebleApp.apk ../vendor/hardware_overlay/TrebleApp/app.apk
+    if [ -f "TrebleApp.apk" ]; then
+        cp TrebleApp.apk ../vendor/hardware_overlay/TrebleApp/app.apk
+    else
+        wget https://raw.githubusercontent.com/phhusson/vendor_hardware_overlay/pie/TrebleApp/app.apk
+        cp app.apk ../vendor/hardware_overlay/TrebleApp/
+    fi
     cd ..
     echo
 }
 
 buildVariant() {
-    echo "--> Building treble_arm64_bvN"
-    lunch treble_arm64_bvN-userdebug
+    echo "--> Building treble_a64_bvN"
+    lunch treble_a64_bvN-userdebug
     make installclean
     make -j$(nproc --all) systemimage
-    mv $OUT/system.img $BD/system-treble_arm64_bvN.img
+    mv $OUT/system.img $BD/system-treble_a64_bvN.img
     echo
 }
 
 buildSlimVariant() {
-    echo "--> Building treble_arm64_bvN-slim"
+    echo "--> Building treble_a64_bvN-slim"
     wget https://gist.github.com/ponces/891139a70ee4fdaf1b1c3aed3a59534e/raw/slim.patch -O /tmp/slim.patch
     (cd vendor/gapps && git am /tmp/slim.patch && rm /tmp/slim.patch)
     make -j$(nproc --all) systemimage
     (cd vendor/gapps && git reset --hard HEAD~1)
-    mv $OUT/system.img $BD/system-treble_arm64_bvN-slim.img
+    mv $OUT/system.img $BD/system-treble_a64_bvN-slim.img
     echo
 }
 
-buildVndkliteVariant() {
-    echo "--> Building treble_arm64_bvN-vndklite"
-    cd sas-creator
-    sudo bash lite-adapter.sh 64 $BD/system-treble_arm64_bvN.img
-    cp s.img $BD/system-treble_arm64_bvN-vndklite.img
-    sudo rm -rf s.img d tmp
-    cd ..
-    echo
-}
 
 generatePackages() {
     echo "--> Generating packages"
-    xz -cv $BD/system-treble_arm64_bvN.img -T0 > $BD/"$BUILD"_arm64-ab-12.1-$BUILD_DATE-UNOFFICIAL.img.xz
-    xz -cv $BD/system-treble_arm64_bvN-vndklite.img -T0 > $BD/"$BUILD"_arm64-ab-vndklite-12.1-$BUILD_DATE-UNOFFICIAL.img.xz
-    xz -cv $BD/system-treble_arm64_bvN-slim.img -T0 > $BD/"$BUILD"_arm64-ab-slim-12.1-$BUILD_DATE-UNOFFICIAL.img.xz
+    xz -cv $BD/system-treble_a64_bvN.img -9 -T0 > $BD/"$BUILD"_a64-ab-12.1-$BUILD_DATE-UNOFFICIAL.img.xz
+    xz -cv $BD/system-treble_a64_bvN-slim.img -9 -T0 > $BD/"$BUILD"_a64-ab-slim-12.1-$BUILD_DATE-UNOFFICIAL.img.xz
     rm -rf $BD/system-*.img
     echo
 }
@@ -119,7 +147,6 @@ setupEnv
 buildTrebleApp
 buildVariant
 buildSlimVariant
-buildVndkliteVariant
 generatePackages
 
 END=`date +%s`
